final code with all changes

+++ src/views/VirtualMachines/VMList.jsx
@@ -20,7 +20,10 @@ import constants from "../../constants";
 import MigrateModal from "./Form/MigrateModal";
 import EditVmModal from "./Form/EditVmModal";
 import { showToastAction } from "../../store/slices/commonSlice";
-import { hasPermission } from "../../utils/commonFunctions";
+import {
+  filterNamespacesByCrudVMS,
+  checkNamespaceValue,
+} from "../../utils/commonFunctions";
 const timeTemplate = (item) => {
   return <>{timeAgo(item.time)}</>;
 };
@@ -50,12 +53,21 @@ const breadcrumItems = [
 ];
 export default function VMList() {
   const dispatch = useDispatch();
-  const { profile } = useSelector((state) => state.user);
-  console.log("profile", profile);
+  const { profile, userNamespace } = useSelector((state) => state.user);
+  let { vms, namespacesDropdown } = useSelector((state) => state.project);
 
-  let { vms } = useSelector((state) => state.project);
-  console.log("vms",vms);
-  
+ 
+
+  const hasAccess = () => {
+    if (profile?.role === "admin") return true;
+    else {
+      const filteredNamespaces = filterNamespacesByCrudVMS(
+        namespacesDropdown,
+        userNamespace
+      );
+      return filteredNamespaces.length > 0;
+    }
+  };
 
   const actionTemplate = (item) => {
     return (
@@ -155,11 +167,17 @@ export default function VMList() {
   }, [dispatch]);
 
   const vmname = (item) => {
-    return (
-      <Link to={`/virtual-machines/details/${item.namespace}/${item.name}`}>
-        {item.name}
-      </Link>
-    );
+
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") || profile?.role === "admin") {
+      return (
+        <Link to={`/virtual-machines/details/${item.namespace}/${item.name}`}>
+          {item.name}
+        </Link>
+      );
+    } else {
+      return <>{item.name}</>;
+    }
+    
   };
   const [visible, setVisible] = useState(false);
   const [editInfo, setEditInfo] = useState(null);
@@ -167,32 +185,63 @@ export default function VMList() {
 
   const ref = useRef();
 
-  const onStart = (item) => {
+  const showError = () => {
     dispatch(
-      onChangeVmStatusAction(item, {
-        running: true,
+      showToastAction({
+        type: "error",
+        title: "Sorry You have no permission!",
       })
     );
+  };
+
+  const onStart = (item) => {
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") ||profile?.role === "admin") {
+      dispatch(
+        onChangeVmStatusAction(item, {
+          running: true,
+        })
+      );
+    } else {
+      showError();
+    }
   };
   const onRestart = (item) => {
-    dispatch(onRestartVMAction(item));
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") ||profile?.role === "admin") {
+      dispatch(onRestartVMAction(item));
+    } else {
+      showError();
+    }
   };
   const onStop = (item) => {
-    dispatch(
-      onChangeVmStatusAction(item, {
-        running: false,
-      })
-    );
+   
+
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") ||profile?.role === "admin") {
+      dispatch(
+        onChangeVmStatusAction(item, {
+          running: false,
+        })
+      );
+    } else {
+      showError();
+    }
   };
   const onPauseUnpause = (item, type) => {
-    dispatch(
-      onPauseVMAction({ ...item, type }, () => {
-        dispatch(getVMsAction());
-      })
-    );
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS")) {
+      dispatch(
+        onPauseVMAction({ ...item, type }, () => {
+          dispatch(getVMsAction());
+        })
+      );
+    } else {
+      showError();
+    }
   };
   const onMigrate = (item) => {
-    setOpenMigrate(item);
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") ||profile?.role === "admin") {
+      setOpenMigrate(item);
+    } else {
+      showError();
+    }
   };
   const onOpenConsole = ({ name, namespace }) => {
     window.open(
@@ -202,20 +251,28 @@ export default function VMList() {
     );
   };
   const onEdit = (item) => {
-    setEditInfo(item);
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") ||profile?.role === "admin") {
+      setEditInfo(item);
+    } else {
+      showError();
+    }
   };
   const onDelete = (item) => {
-    confirmDialog({
-      target: ref.currentTarget,
-      header: "Delete Confirmation",
-      message: `Do you want to delete ${item.namespace} - ${item.name} ?`,
-      icon: "pi pi-info-circle",
-      rejectClassName: "p-button-outlined p-button-secondary",
-      acceptClassName: " primary-button",
-      accept: () => {
-        dispatch(onDeleteVMAction(item));
-      },
-    });
+    if (checkNamespaceValue(userNamespace, item.namespace, "crudVMS") ||profile?.role === "admin") {
+      confirmDialog({
+        target: ref.currentTarget,
+        header: "Delete Confirmation",
+        message: `Do you want to delete ${item.namespace} - ${item.name} ?`,
+        icon: "pi pi-info-circle",
+        rejectClassName: "p-button-outlined p-button-secondary",
+        acceptClassName: " primary-button",
+        accept: () => {
+          dispatch(onDeleteVMAction(item));
+        },
+      });
+    } else {
+      showError();
+    }
   };
 
   const [search, setSearch] = useState("");
@@ -232,8 +289,8 @@ export default function VMList() {
     if (profile.role === "admin") {
       setVisible(true);
     } else {
-      if (hasPermission(profile, "vm-create")) {
-
+      if (hasAccess) {
+        setVisible(true);
       } else {
         dispatch(
           showToastAction({
